Made DB and schema roles part of DB API
