+++ mathesar/urls.py
@@ -22,13 +22,17 @@ db_table_router.register(r'constraints', db_viewsets.ConstraintViewSet, basename
 ui_router = routers.DefaultRouter()
 ui_router.register(r'databases', ui_viewsets.DatabaseViewSet, basename='database')
 ui_router.register(r'users', ui_viewsets.UserViewSet, basename='user')
-ui_router.register(r'database_roles', ui_viewsets.DatabaseRoleViewSet, basename='database_role')
-ui_router.register(r'schema_roles', ui_viewsets.SchemaRoleViewSet, basename='schema_role')
+
+ui_database_router = routers.NestedSimpleRouter(ui_router, r'databases', lookup='database')
+ui_database_router.register(r'database_roles', ui_viewsets.DatabaseRoleViewSet, basename='database-database-role')
+ui_database_router.register(r'schema_roles', ui_viewsets.SchemaRoleViewSet, basename='database-schema-role')
+
 
 urlpatterns = [
     path('api/db/v0/', include(db_router.urls)),
     path('api/db/v0/', include(db_table_router.urls)),
     path('api/ui/v0/', include(ui_router.urls)),
+    path('api/ui/v0/', include(ui_database_router.urls)),
     path('api/ui/v0/reflect/', views.reflect_all, name='reflect_all'),
     path('auth/', include('django.contrib.auth.urls')),
